Braintree: Support override_application_id
COMP-266

ActiveMerchant currently uses Class.application_id or @options[:channel] to let merchants
pass platform BN codes to gateways to signify that a transaction may be going to one merchant's
accounts but it originates from a platform or aggregator.

This poses a problem in that the Class value is assigned at compile time and the @options[:channel]
refers to a static value used when initializing the class.

This commit adds support for `override_application_id` which is a way for aggregators to pass in a
new BN code at transaction time alongside transaction parameters.

Test Summary
Remote:
123 tests, 661 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed

Cannot add a remote test because there is no generic `channel` value